+++ tests/test_config.py
@@ -20,6 +20,27 @@ run_config:
     start_date: 20200102
     image_pull_secrets: pull1,pull2
     service_account_name: service_account
+    failure_handlers:
+      - type: slack
+        connection_id: slack
+        message_template: "test message"
+    spark:
+      type: dataproc
+      region: europe-west2
+      project_id: sandbox
+      cluster_name: cluster-8084
+      artifacts_path: gs://test/dataproc
+      cluster_config:
+        configBucket: test
+    kubernetes_pod_templates:
+        custom_template:
+            template: |-
+                type: Pod
+                metadata:
+                    name: {PodGenerator.make_unique_pod_id('test')}
+                spec:
+                    name: 1
+            image: custom
     volume:
         storageclass: kms
         size: 3Gi
@@ -80,8 +101,8 @@ run_config:
 
 class TestPluginConfig(unittest.TestCase):
     def test_plugin_config(self):
-        cfg = PluginConfig(yaml.safe_load(CONFIG_YAML))
 
+        cfg = PluginConfig(yaml.safe_load(CONFIG_YAML))
         assert cfg.host == "test.host.com"
         assert cfg.output == "/data/ariflow/dags"
         assert cfg.run_config
@@ -104,6 +125,10 @@ class TestPluginConfig(unittest.TestCase):
         assert cfg.run_config.volume.owner == 1000
         assert cfg.run_config.volume.disabled is True
         assert cfg.run_config.resources
+        failure_handler = cfg.run_config.failure_handlers[0]
+        assert failure_handler.type == "slack"
+        assert failure_handler.connection_id == "slack"
+        assert failure_handler.message_template == """test message"""
         resources = cfg.run_config.resources
         assert resources.__default__
         assert resources.__default__.node_selectors
@@ -116,6 +141,7 @@ class TestPluginConfig(unittest.TestCase):
             "value": "data-processing",
             "effect": "NoExecute",
         }
+
         assert resources.__default__.annotations
         assert (
             resources.__default__.annotations["iam.amazonaws.com/role"]
@@ -191,6 +217,27 @@ postgres://{{ .Data.username }}:{{ .Data.password }}@postgres:5432/mydb
             "MLFLOW_TRACKING_PASSWORD",
         ]
 
+        kubernetes_pod_templates = cfg.run_config.kubernetes_pod_templates
+
+        assert kubernetes_pod_templates.custom_template.image == "custom"
+        assert (
+            kubernetes_pod_templates.custom_template.template
+            == """type: Pod
+metadata:
+    name: {PodGenerator.make_unique_pod_id('test')}
+spec:
+    name: 1"""
+        )
+
+        spark = cfg.run_config.spark
+        assert spark
+        assert spark.type == "dataproc"
+        assert spark.region == "europe-west2"
+        assert spark.project_id == "sandbox"
+        assert spark.cluster_name == "cluster-8084"
+        assert spark.artifacts_path == "gs://test/dataproc"
+        assert spark.cluster_config["configBucket"] == "test"
+
     def test_defaults(self):
         cfg = PluginConfig({"run_config": {}})
 
@@ -202,6 +249,7 @@ postgres://{{ .Data.username }}:{{ .Data.password }}@postgres:5432/mydb
         assert cfg.run_config.start_date is None
         assert cfg.run_config.auth_config.type == "Null"
         assert cfg.run_config.auth_config.params == []
+        assert cfg.run_config.spark
 
         assert cfg.run_config.volume
         assert cfg.run_config.volume.disabled is False
@@ -213,6 +261,8 @@ postgres://{{ .Data.username }}:{{ .Data.password }}@postgres:5432/mydb
         assert cfg.run_config.resources
 
         assert not cfg.run_config.external_dependencies
+
+        assert len(cfg.run_config.kubernetes_pod_templates) == 0
 
     def test_run_name_is_experiment_name_by_default(self):
         cfg = PluginConfig(